+++ mitosheet/dev/metaprogramming/create_new_step.py
@@ -99,7 +99,7 @@ from mitosheet.code_chunks.code_chunk import CodeChunk
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import ColumnID
 
 class {step_performer_name}(StepPerformer):
