+++ mitosheet/mitosheet/state.py
@@ -12,7 +12,7 @@ from mitosheet.column_headers import ColumnIDMap
 from mitosheet.types import FrontendFormulaAndLocation
 from mitosheet.types import ColumnHeader, ColumnID, DataframeFormat
 from mitosheet.user.utils import is_enterprise, is_running_test
-from mitosheet.utils import check_valid_sheet_functions, get_first_unused_dataframe_name
+from mitosheet.utils import  get_first_unused_dataframe_name
 
 # Constants for where the dataframe in the state came from
 DATAFRAME_SOURCE_PASSED = "passed"  # passed in mitosheet.sheet
