+++ mitosheet/mitosheet/step_performers/column_headers_transform.py
@@ -13,7 +13,7 @@ from mitosheet.errors import make_column_exists_error
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import ColumnID
 
 class ColumnHeadersTransformStepPerformer(StepPerformer):
