+++ mitosheet/mitosheet/step_performers/column_steps/delete_column.py
@@ -12,7 +12,7 @@ from mitosheet.code_chunks.step_performers.column_steps.delete_column_code_chunk
 from mitosheet.errors import raise_error_if_column_ids_do_not_exist
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import ColumnID
