+++ mitosheet/mitosheet/step_performers/dataframe_steps/dataframe_delete.py
@@ -10,7 +10,7 @@ from mitosheet.code_chunks.step_performers.dataframe_steps.dataframe_delete_code
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 
 
 class DataframeDeleteStepPerformer(StepPerformer):
