+++ mitosheet/mitosheet/step_performers/delete_row.py
@@ -12,7 +12,7 @@ from mitosheet.code_chunks.delete_row_code_chunk import DeleteRowCodeChunk
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import ColumnID
 
 class DeleteRowStepPerformer(StepPerformer):
