+++ mitosheet/mitosheet/step_performers/export_to_file.py
@@ -15,7 +15,7 @@ import os
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 
 from mitosheet.utils import write_to_excel
