+++ mitosheet/mitosheet/step_performers/graph_steps/graph_duplicate.py
@@ -11,7 +11,7 @@ from mitosheet.code_chunks.empty_code_chunk import EmptyCodeChunk
 
 from mitosheet.state import State
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import GraphID
