+++ mitosheet/mitosheet/step_performers/promote_row_to_header.py
@@ -16,7 +16,7 @@ from mitosheet.errors import make_invalid_promote_row_to_header
 from mitosheet.state import State
 from mitosheet.step_performers.column_steps.rename_column import rename_column_headers_in_state
 from mitosheet.step_performers.step_performer import StepPerformer
-from mitosheet.step_performers.utils import get_param
+from mitosheet.step_performers.utils.utils import get_param
 from mitosheet.types import ColumnHeader
 from mitosheet.utils import MAX_ROWS, convert_df_to_parsed_json
 from mitosheet.public.v3 import deduplicate_column_headers
